test: `suspend()` then `notify()`
