+++ crates/moot/tests/moot_lmoo.rs
@@ -5,18 +5,12 @@
 //! * MOOT_PORT: port the MOO server listens on, defaults to 7777
 
 use std::{
-    collections::HashMap,
     env,
-    fs::File,
-    io::{BufRead, BufReader},
     path::{Path, PathBuf},
     process::{Command, Stdio},
-    time::{Duration, Instant},
 };
 
-use eyre::Context;
-use moor_moot::{test_db_path, ManagedChild, MootClient, MootRunner, MootState, WIZARD};
-use moor_values::var::Objid;
+use moor_moot::{execute_moot_test, test_db_path, ManagedChild, TelnetMootRunner};
 
 fn moo_path() -> PathBuf {
     env::var("MOOT_MOO_PATH")
@@ -53,90 +47,9 @@ fn start_moo() -> ManagedChild {
     )
 }
 
-struct TelnetMootRunner {
-    clients: HashMap<Objid, MootClient>,
-}
-impl TelnetMootRunner {
-    fn new() -> Self {
-        Self {
-            clients: HashMap::new(),
-        }
-    }
-
-    fn client(&mut self, player: Objid) -> &mut MootClient {
-        self.clients.entry(player).or_insert_with(|| {
-            let start = Instant::now();
-            loop {
-                if let Ok(mut client) = MootClient::new(moo_port()) {
-                    client
-                        .send_string(std::format!("connect {}", player))
-                        .unwrap();
-                    return client;
-                } else if start.elapsed() > Duration::from_secs(5) {
-                    panic!("Failed to connect to daemon");
-                } else {
-                    std::thread::sleep(Duration::from_millis(10));
-                }
-            }
-        })
-    }
-
-    fn resolve_response(&mut self, response: String) -> Result<String, std::io::Error> {
-        // Resolve the response; for example, the test assertion may be `$object`; resolve it to the object's specific number.
-        self.client(WIZARD).command(format!(
-            "; return {response}; \"TelnetMootRunner::resolve_response\";"
-        ))
-    }
-}
-impl MootRunner for TelnetMootRunner {
-    type Value = String;
-    type Error = std::io::Error;
-
-    fn eval<S: Into<String>>(
-        &mut self,
-        player: Objid,
-        command: S,
-    ) -> Result<String, std::io::Error> {
-        let response = self
-            .client(player)
-            .command(format!("; {} \"TelnetMootRunner::eval\";", command.into()))?;
-        self.resolve_response(response)
-    }
-
-    fn command<S: AsRef<str>>(
-        &mut self,
-        player: Objid,
-        command: S,
-    ) -> Result<String, std::io::Error> {
-        let response = self.client(player).command(command)?;
-        self.resolve_response(response)
-    }
-
-    fn none(&self) -> Self::Value {
-        "0".to_string()
-    }
-}
-
 fn test_moo(path: &Path) {
     let mut _moo = start_moo();
-
-    let f = BufReader::new(
-        File::open(path)
-            .wrap_err(format!("{}", path.display()))
-            .unwrap(),
-    );
-
-    let mut state = MootState::new(TelnetMootRunner::new(), WIZARD);
-    for (line_no, line) in f.lines().enumerate() {
-        let line = line.unwrap();
-        let line_no = line_no + 1;
-        state = state
-            .process_line(line_no, &line)
-            .wrap_err(format!("line {}", line_no))
-            .unwrap();
-        //eprintln!("[{line_no}] {line} {state:?}");
-    }
-    state.finalize().unwrap();
+    execute_moot_test(TelnetMootRunner::new(moo_port()), path)
 }
 
 #[test]
