+++ crates/telnet-host/Cargo.toml
@@ -41,5 +41,4 @@ uuid.workspace = true
 # Testing
 [dev-dependencies]
 escargot.workspace = true
-pretty_assertions.workspace = true
 tempfile.workspace = true
