+++ crates/telnet-host/tests/moot/echo.moot
@@ -1,0 +1,2 @@
+; return 42;
+42
