+++ crates/telnet-host/tests/moot/suspend_notify.moot
@@ -1,0 +1,3 @@
+; suspend(1); notify(player, "200"); return 201;
+200
+201
